Tweak metadata provider, add tests
Since the metadata provider already uses an environment variable to
determine which host it should call, we can use the httptest package's
test server to simulate our metadata endpoint in tests. This removes the
need to have an extra interface to shim the HTTP client, and lets the
metadata provider just use it directly.